Rename API file to the original name
