+++ src/libs/expensifyAPI.js
@@ -1,7 +1,7 @@
 import Onyx from 'react-native-onyx';
 import _ from 'underscore';
 import Network from './Network';
-import API from './API2';
+import API from './API';
 import ONYXKEYS from '../ONYXKEYS';
 import redirectToSignIn from './actions/SignInRedirect';
 import CONFIG from '../CONFIG';
